requirements.txt
mcp[server-fastmcp]
google-genai
python-dotenv
pydantic
rich
pymongo
clickhouse-connect
requests
urllib3